First Pr fixes
