+++ client/QuickStart/kusto_sample_config.json
@@ -1,19 +1,18 @@
 {
-	"kustoUri" : "https://kvc8145132f55d24729976.northeurope.kusto.windows.net",
-	"ingestUri" : "https://ingest-kvc8145132f55d24729976.northeurope.kusto.windows.net",
-	"databaseName" : "MyDatabase",
-	"tableName" : "SampleTable",
+	"kustoUri": "https://kvc8145132f55d24729976.northeurope.kusto.windows.net",
+	"ingestUri": "https://ingest-kvc8145132f55d24729976.northeurope.kusto.windows.net",
+	"databaseName": "MyDatabase",
+	"tableName": "SampleTable",
 	"useExistingTable": true,
 	"alterTable": true,
 	"queryData": true,
 	"ingestData": true,
-	"AuthenticationMode" :"UserPrompt",
+	"AuthenticationMode": "UserPrompt",
 	"WaitForUser": true,
-	"WaitForIngestSeconds" : 20,
-	"BatchingPolicy" : "{ 'MaximumBatchingTimeSpan': '00:00:10', 'MaximumNumberOfItems': 500, 'MaximumRawDataSizeMB': 1024 }",
-	"tableSchema" : "(rownumber:int, rowguid:string, xdouble:real, xfloat:real, xbool:bool, xint16:int, xint32:int, xint64:long, xuint8:long, xuint16:long, xuint32:long, xuint64:long, xdate:datetime, xsmalltext:string, xtext:string, xnumberAsText:string, xtime:timespan, xtextWithNulls:string, xdynamicWithNulls:dynamic)",
-	"data" :
-	[
+	"WaitForIngestSeconds": 20,
+	"BatchingPolicy": "{ 'MaximumBatchingTimeSpan': '00:00:10', 'MaximumNumberOfItems': 500, 'MaximumRawDataSizeMB': 1024 }",
+	"tableSchema": "(rownumber:int, rowguid:string, xdouble:real, xfloat:real, xbool:bool, xint16:int, xint32:int, xint64:long, xuint8:long, xuint16:long, xuint32:long, xuint64:long, xdate:datetime, xsmalltext:string, xtext:string, xnumberAsText:string, xtime:timespan, xtextWithNulls:string, xdynamicWithNulls:dynamic)",
+	"data": [
 		{
 			"sourceType": "localFileSource",
 			"dataSourceUri": "dataset.csv",
